extract server.js from tests/test-web.js
